Inserida a funcionalidade de editar uma refeição e retornando a refeição editada no dashboard.
